% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{age_structure}
\alias{age_structure}
\title{Set-up age groups and find ageing rate
Takes the global variables numMonthly and final_age, along
With the appropriate greater Perth data file. Assuming a non-uniform,
but constant population distribution}
\usage{
age_structure(year, area = 0)
}
\arguments{
\item{year}{ABS greater Perth population year}

\item{area}{0 or 1. Default is 0 = "Greater Perth", 1 = "Southern WA"}
}
\value{
list nAges, pop_groups, age_in, age_out
}
\description{
Set-up age groups and find ageing rate
Takes the global variables numMonthly and final_age, along
With the appropriate greater Perth data file. Assuming a non-uniform,
but constant population distribution
}
