% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{aggregate_output_immi}
\alias{aggregate_output_immi}
\title{TO DO: CHANGE TO PASSING nState and nRet COULD THEN POTENTIALLY COLLAPSE THIS
FUNCTION AND ABOVE INTO ONE
Takes output from deSolve and pulls out hospitalisations aggregated in the
age groups used for fitting for term first exposure, term subsequent exposures,
term combined, preterm first exposure, preterm subsequent exposures,
preterm combined, combined first exposure, combined second exposure, combined}
\usage{
aggregate_output_immi(deSolveOut, times, nAges, nState, nRet, old = 0)
}
\arguments{
\item{deSolveOut}{output matrix}

\item{times}{time vector}

\item{nAges}{number of age groups}

\item{old}{1 = include older age group 24-<60 months,
2 = include 26-<36 months, 36->48 months, 48-<60 months}

\item{numImm}{number of immunisation categories}
}
\value{
list
}
\description{
TO DO: CHANGE TO PASSING nState and nRet COULD THEN POTENTIALLY COLLAPSE THIS
FUNCTION AND ABOVE INTO ONE
Takes output from deSolve and pulls out hospitalisations aggregated in the
age groups used for fitting for term first exposure, term subsequent exposures,
term combined, preterm first exposure, preterm subsequent exposures,
preterm combined, combined first exposure, combined second exposure, combined
}
