% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_likelihood_base_new}
\alias{create_likelihood_base_new}
\title{Set up likelihood function for base model in a closure environment
for use with lazymcmc - this likelihood includes the assumption that most
children by the age of 2 year old have been infected}
\usage{
create_likelihood_base_new(parTab, data, fixedPar, PRIOR_FUNC, ...)
}
\arguments{
\item{parTab}{parameter table setup for use in lazymcmc fitting functions}

\item{data}{observational incidence data}

\item{fixedPar}{fixed parameter values}

\item{PRIOR_FUNC}{optional prior function}
}
\value{
function
}
\description{
Set up likelihood function for base model in a closure environment
for use with lazymcmc - this likelihood includes the assumption that most
children by the age of 2 year old have been infected
}
