% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.contactDaily.ABSMicro}
\alias{data.contactDaily.ABSMicro}
\title{Daily contact matrix}
\format{
A 75 x 75 data matrix
}
\source{
\url{https://rdrr.io/github/njtierney/conmat/}
}
\usage{
data.contactDaily.ABSMicro
}
\description{
Daily number of contacts between age groups within the population.
This contact matrix was generated using the R package conmat based on the
POLYMOD study results for the UK, normalised to 2016 metropolitan Perth
population demographics but with finer age-stratification of mixing between
household contacts based on the demographics of the 2016 ABS 5\% microdata
sample for the Greater Perth region and between school/childcare contacts of
children under 5 based on the Childhood Education and Care Survey 2017.
The <5 year age group is divided into monthly age groups in the model,
with contacts being uniformly distributed into these resulting age groups.
}
\keyword{datasets}
