% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.southernWA.2016}
\alias{data.southernWA.2016}
\title{ABS population of Greater Perth by 5-year age group - 2016}
\format{
A data table containing 18 observations of 4 variables.
\describe{
\item{lga}{Local Government Area classification}
\item{lower.age.limit}{}
\item{year}{census data year}
\item{population}{population in age group}
}
}
\source{
\url{https://abs.gov.au/census/find-census-data/quickstats/2021/5GPER}
}
\usage{
data.southernWA.2016
}
\description{
Data is formatted to be easily used with the R package conmat
to find the applicable contact matrix
}
\details{
Children in Kimberley & Pilbara regions identified via the following postcodes were excluded:
\itemize{
\item Kimberley: 6725, 6726, 6728, 6731, 6733, 6740, 6743, 6765, 6770
\item Pilbara: 6710, 6711, 6712, 6713, 6714, 6715, 6716, 6718, 6720, 6721,
}
}
\keyword{datasets}
