% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findPhaseShift}
\alias{findPhaseShift}
\title{Find phase shift relating to a set "peak" month, default is 7 (July)}
\usage{
findPhaseShift(counts, month = 7)
}
\arguments{
\item{counts}{table of modelled detected infections in aggregated age groups,
when phi = 0}

\item{month}{month 1 - 12, Default is 7 (July)}
}
\value{
numeric phase
}
\description{
Find phase shift relating to a set "peak" month, default is 7 (July)
}
